Disconnect client from created/dropped database
This allows subsequent code (using other libraries that depend on a database connection) to connect to an up-to-date database. For example, some code that creates tables in the new database. This is the same changes as seen here https://github.com/olalonde/pgtools/pull/20/commits/a2579da4ca60b41b023ea311649a8237f5fe01f8 which is a stale MR at this time.